+++ src/main/java/dataDouble/Ejercicio3.java
@@ -1,18 +1,14 @@
-package org.example;
+package dataDouble;
 
 import java.util.Random;
 
-public class Ejercicio4 {
-
+public class Ejercicio3 {
     public static void main(String[] args) {
         double[] miArray = generator();
 
-        ordenarArray(miArray);
+        System.out.println("La suma es " + suma(miArray));
 
 
-        for (double valor : miArray) {
-            System.out.println(valor + "     ");
-        }
     }
 
     public static double[] generator() {
@@ -33,19 +29,14 @@ public class Ejercicio4 {
         return arrayAleatorio;
     }
 
-
-    public static void ordenarArray(double[] arr) {
-        //Comentario de prueba
-        int n = arr.length;
-        double tempo = 0;
-        for (int i = 0; i < n - 1; i++) {
-            for (int j = 0; j < n - i - 1; j++) {
-                if (arr[j] < arr[j + 1]) {
-                    tempo = arr[j];
-                    arr[j] = arr[j + 1];
-                    arr[j + 1] = tempo;
-                }
-            }
+    public static double suma(double[] array) {
+        double sumav = 0;
+        for (double elemento : array) {
+            sumav += elemento;
         }
+
+
+        return sumav;
     }
+
 }
